requirements.txt
streamlit>=1.24.0
arxiv
requests
python-dotenv
boto3
markdown
pandas
python-dateutil
beautifulsoup4
lxml